int long problem

+++ Kinoheld.Api.Client/Kinoheld.Api.Client/Api/IKinoheldApiClient.cs
@@ -10,7 +10,7 @@ namespace Kinoheld.Api.Client.Api
     {
         Task<JObject> GetCinemas(string city, string searchTerm, int distance, GetCinemasDynamicQuery dynamicQuery, CancellationToken cancellationToken);
 
-        Task<JObject> GetShows(int cinemaId, DateTime? date, GetShowsDynamicQuery dynamicQuery, CancellationToken cancellationToken);
+        Task<JObject> GetShows(long cinemaId, DateTime? date, GetShowsDynamicQuery dynamicQuery, CancellationToken cancellationToken);
 
         Task<JObject> GetCities(string searchTerm, int limit, CancellationToken cancellationToken);
     }
